[ADD]saas_portal_more_like_this:
Module allows you to provision a client instance that is an exact replica of
another existing client.

The reason for this being that it is not uncommon for clients instance to evolve
far beyond the template on which they are based. And so this allows you to
provision an identical instance in seconds rather than have to re-appply all your
changes to the template.

This module is very useful in cases in which you wish to provide a training
environment for a client based on the exact copy of their live instance